src/test/java/com/clholdings/pass_shoot/springboot/web/ProfileControllerUnitTest.java
package com.clholdings.pass_shoot.springboot.web;

import org.junit.jupiter.api.Test;
import org.springframework.mock.env.MockEnvironment;

import static org.assertj.core.api.Assertions.assertThat;

public class ProfileControllerUnitTest {

    @Test
    public void real_profile이_조회된다() {
        //given
        String expectedProfile = "real";
        MockEnvironment env = new MockEnvironment();
        env.addActiveProfile(expectedProfile);
        env.addActiveProfile("oauth");
        env.addActiveProfile("real-db");

        // Environment는 인터페이스라 가짜 구현체인 MockEnvironment(스프링에서 제공)을 사용해서 테스트
        // 하면됩니다
        // 이렇게 해보면 생성자 DI가 얼마나 유용한지알수있습니다. 만약 Environment를 @Autowired로
        // 받았다면 이런 테스트코드를작성하지 못했습니다.항상 스프링 테스트를 해야만 했겠죠? 앞의
        // 테스트가 성공적으로 다 통과했다면 컨트롤러 로직에 대한 이슈는 없습니다
        ProfileController controller = new ProfileController(env);

        //when
        String profile = controller.profile();

        //then
        assertThat(profile).isEqualTo(expectedProfile);
    }

    @Test
    public void real_profile이_없으면_첫번째가_조회된다() {
        //given
        String expectedProfile = "oauth";;
        MockEnvironment env = new MockEnvironment();

        env.addActiveProfile(expectedProfile);
        env.addActiveProfile("real-db");

        ProfileController controller = new ProfileController(env);

        //when
        String profile = controller.profile();

        //then
        assertThat(profile).isEqualTo(expectedProfile);
    }

    @Test
    public void active_profile이_없으면_default가_조회된다() {
        //given
        String expectedProfile = "default";
        MockEnvironment env = new MockEnvironment();
        ProfileController controller = new ProfileController(env);

        //when
        String profile = controller.profile();

        //then
        assertThat(profile).isEqualTo(expectedProfile);
    }
}
